Fix for issue 4284 (Add ability to Duplicate an existing entry with the Control Panel).
Add a new menu item in both the 'Entries' and popup menu of the 'Manage Entries' dialog.  This item is labeled 'Duplicate Entry...'.

The user must provide the new name of the entry and the new parent DN for the entry.
There is a limitation to duplicate the entries with a password.  If the original entry has a password, the resulting duplicated entry will have 'password' as password value.  The user is informed of this in the dialog.

Some improvements have been made in the layout used in StatusGenericPanel to display messages an errors.

A fix has been done in the focus management of GenericDialog.


git-svn-id: https://svn.forgerock.org/opendj/trunk@6008 41b1ffd8-f28e-4786-ab96-9950f0a78031