Fix for issue 4310 (dsreplication status - Unexpected error reading replica ID: NumberFormatException)
Use the logger of the control panel that disables the logging earlier.

Fix lots of error codes in dsreplication (most of the commands used to return SUCCESSFUL_NOP when they should return SUCCESSFUL).

Fix some warnings reported by the IDE by using generics.

Now the log file generated by dsreplication is shown to the user for some sub-commands even in the case where the operation was successful.


git-svn-id: https://svn.forgerock.org/opendj/trunk@6019 41b1ffd8-f28e-4786-ab96-9950f0a78031