Fix for issue 4326 (Allow setup of a Server without Data suffix)
If the user does not specify the option baseDN (in the command-line setup) or leaves the base DN text field empty (in the graphical setup), no b
ase DN will be created.

Fix some issues with the installer, that on Mac OS used to make the terminal to lose focus before the confirmation of the setup.



git-svn-id: https://svn.forgerock.org/opendj/trunk@6075 41b1ffd8-f28e-4786-ab96-9950f0a78031