Commit a potential fix for issue 4394 (Quicksetup doesn't work with openJDK 1.6 64 bits (fedora core 12))
The fix is not really a fix for this, the code silently ignored the non existence of the javax.jnlp.DownloadService.  The modifications consist basically on identifying this situation and providing an error message informing of this and of the alternatives the user has to setup OpenDS (download a ZIP file).


git-svn-id: https://svn.forgerock.org/opendj/trunk@6264 41b1ffd8-f28e-4786-ab96-9950f0a78031