Fix for issue 3901 (Setup fails when inserting a server in a replication topology with large umount of data) - Add a 'Runtime Options' step in the wizard.  By default this will contain just two labels and two buttons: Server Runtime Options: Use Default <button to change> Import Options: Use Default <button to change>
If the user clicks on one of these buttons, (s)he will be able to modify the runtime settings for the import and the server.

There are also some messages trying to help the user.The idea is to find out if the user wanted to import a big LDIF file, a lot of automatically generated data or if the replicated base DNs contain lot of entries.  If it is the case and we consider that the default memory we allow to run the server (256 Mb) is not enough, the panel will display a message informing that it is recommended to increase the memory required.   For the moment the message is just a warning and has not a lot of information about a recommended memory size, but it improves the current behavior, were the import fails because of a lack of heap in the JVM.



git-svn-id: https://svn.forgerock.org/opendj/trunk@6358 41b1ffd8-f28e-4786-ab96-9950f0a78031