Problem pointed out by Alin Brici.
This link is permanently broken, except in the finished docs (post release). So it merits a bit of explanation.

If I were to point in these (pre-2.5.0) docs to the 2.4.4 QuickSetup, then that would be wrong, because the docs are aimed at 2.5.0. The currently correct link is the nightly build, but that's http://download.forgerock.org/downloads/opendj/20111123023002/install/QuickSetup.jnlp and will change tomorrow.

git-svn-id: https://svn.forgerock.org/opendj/trunk@7499 41b1ffd8-f28e-4786-ab96-9950f0a78031