Fix OPENDJ-435: Don't create any backend DB's when running setup in cli mode
Remove the default value of "dc=example,dc=com" for the -b option to setup, thereby allowing non-interactive creation of a new OpenDJ instance without a backend.

Note that both interactive CLI and GUI setup interfaces permit creation of a new instance without a backend, it was just non-interactive mode which had the limitation.

git-svn-id: https://svn.forgerock.org/opendj/trunk@7800 41b1ffd8-f28e-4786-ab96-9950f0a78031