Initial fix for OPENDJ-249: dsreplication disable --disableAll error removing contents of "cn=admin data"
Disabling replication now avoids the self-disconnect by not deleting the administrator entry, so the command now completes successfully. However, it is erasing the entire contents of the ads truststore, including the server instance key. This is definitely not a good idea since the key may be required for reversible encryption.


git-svn-id: https://svn.forgerock.org/opendj/trunk@7828 41b1ffd8-f28e-4786-ab96-9950f0a78031