Final fix for OPENDJ-249: dsreplication disable --disableAll error removing contents of "cn=admin data"
Don't remove instance keys when disabling replication. Instance keys are intended for other purposes as well as replication, for example log signing, etc, and may be required after an instance has been removed from a topology. It is the responsibility of administrators to manually remove unused instance keys on a case by case basis.

git-svn-id: https://svn.forgerock.org/opendj/trunk@7829 41b1ffd8-f28e-4786-ab96-9950f0a78031