OPENDJ-655 Message about authentication failures should contain identification of the user for easier analysis.
What a convoluted code!! It is using a label and a block to implement a goto in Java. yuck!

LocalBackendBindOperation.java:
In processLocalBind(), extracted method processBind().
In the newly extracted method processBind(), converted the break stmts into return stmts + moved the switch stmt inside the try stmt (instead of the try/catch stmt in the switch stmt) + then removed useless if statements and break statements from the switch stmt.


git-svn-id: https://svn.forgerock.org/opendj/trunk@8555 41b1ffd8-f28e-4786-ab96-9950f0a78031