OPENDJ-808 Implement a simple commons REST based HTTP connection handler
ConcurrentHashMap does not accept putting in null values, so I will put in the key itself (It is redundant, but this is not a problem here.)

HTTPConnectionHandler.java:
clientConnections changed from Map<ClientConnection, Void> to Map<ClientConnection, ClientConnection>.

CollectClientConnectionsFilter.java:
Consequence  of the change to HTTPConnectionHandler. 
Use the key to put in the value.

git-svn-id: https://svn.forgerock.org/opendj/trunk@8622 41b1ffd8-f28e-4786-ab96-9950f0a78031