OPENDJ-830 (CR-1595) Implement authentication and authorization for HTTP connection handler
Fixed information leak on unsuccessful authentication: if the username was incorrect, the server was returning 404 (not found) instead of 401 (invalid credentials).
When authentication was unsuccessful, the client was not receiving a response and the connection was not closed leaving it hanging.


CollectClientConnectionFilter.java:
In DoBindResultHandler.handleErrorResult(), send back an authentication failure when a search on the user name does not return a unique result.
In sendAuthenticationFailure(), correctly complete the AsyncContext.
In doFilter(), removed the call to AsyncContext.setTimeout().

git-svn-id: https://svn.forgerock.org/opendj/trunk@8784 41b1ffd8-f28e-4786-ab96-9950f0a78031