Made code more generic: Used Map instead of HashMap in method signatures.
AttrHistoricalMultiple.java, AttrHistoricalSingle.java:
Changed comments to javadocs.
Fixed condition checking.
Inlined local variables and used the field instead.
Replaced use of Boolean with boolean.

We were looking at this code with Ludo and thought replacing the List with a Map had to be done in conjunction with breaking up AttrValueHistorical class to reduce memory usage (removing stroing unnecessary references).
The value field should be used as the key of the Map, and the value of the Map should contain an object with a reference to a ChangeNumber and a boolean indicating whether it is an update or a delete (Since they are mutually exclusive).

git-svn-id: https://svn.forgerock.org/opendj/trunk@8813 41b1ffd8-f28e-4786-ab96-9950f0a78031