OPENDJ-844 Patches don't affect ControlPanel.app OPENDJ-842 On Windows, the setup command hangs when the length of the install path is too long
Code review and initial idea: violette


Now that the application starts with the single bootstrap jar, MacOS X can take advantage of this by referencing it statically in the plist files. In other words, it is no longer necessary to build the plist files dynamically.


Info.plist:
Hardcoded the path to bootstrap.jar

build.xml:
Removed the code that appended the jars to the **/Info.plist files for MacOS X.

git-svn-id: https://svn.forgerock.org/opendj/trunk@8897 41b1ffd8-f28e-4786-ab96-9950f0a78031