OPENDJ-942 (CR-1744) HTTP Connection Handler - Running an authenticated request at "/" hangs
The problem was due to a misconfiguration of Grizzly HTTP server.
Using HttpServer.createSimpleServer() to create the server means Grizzly adds unwanted network listeners and HTTP handlers that interfere with the expected behaviour of the application.
Problem was solved by doing a direct configuration of the HTTP server, removing unwanted mappings and servlets.


HTTPConnectionHandler.java:
In startHttpServer(), extracted methods createHttpServer() and createAndRegisterServlet().
In createHttpServer(), do not use HttpServer.createSimpleServer() anymore and do a direct, clutterless configuration.
In setHttpStatsProbe() and getHttpConfig(), added an HTTPServer parameter so this method can be used before or after creating the HTTPServer.

git-svn-id: https://svn.forgerock.org/opendj/trunk@8917 41b1ffd8-f28e-4786-ab96-9950f0a78031