Fix OPENDJ-176 - Use SLF4J for logging in the SDK Review CR-2404
SDK now only depends on SLF4J API for logging (slf4j-api-1.7.5.jar).
End-users of the SDK are now able to choose the actual logging framework to use by adding an additional dependency. 

For tests only (test scope in maven), there is a also a dependency on java.util.logging (known as jdk14 by SLF4J), in order to change the defaut log level for some test cases. 

There are now multiples loggers, defined centrally in StaticUtils class : DEFAULT, CONTROLS, SCHEMA and IO. 
It gives a central point for the list of available loggers. 
This is a compromise between having a single logger for all messages and one logger per class which can be a hassle to setup. 

The majority of "if (logger.isEnabledForXXX)" checks before logging have been removed as it is the preferred way for SLF4J. 
They are kept only in some cases where arguments of the logged message involve some computation. 

All logs at error, warn and info level should be internationalised, but it will be done in a separate phase (added "TODO: I18N" comments to find them).

git-svn-id: https://svn.forgerock.org/opendj/trunk@9644 41b1ffd8-f28e-4786-ab96-9950f0a78031