Maybe the end of endless pain and suffering on the Continuous Integration front.
In ExternalChangeLogTest.ECLReplicationServerFullTest(), it looks like the server does not have the time to trim the last "changetype: delete" from the previous tests method, so let's force the test to clear the changelogDB before it finishes.

ExternalChangeLogTest.java:
In ECLReplicationServerFullTest(), called clearChangelogDB() in the finally block to force cleaning test data.

git-svn-id: https://svn.forgerock.org/opendj/trunk@9645 41b1ffd8-f28e-4786-ab96-9950f0a78031