OPENDJ-346 Consider using java.util.ServiceLoader for loading extensions and requesting transport implementations
Fixed an issue in StaticUtils#getProvider : passing a null class loader to ServiceLoader#load(Class<T>, ClassLoader) has unexpected effect, 
ie it is not throwing a NPE and has not the same behavior that ServiceLoader#load(Class<T>).

Now code explicitely checks for null class loader and use the appropriate load method.

As a consequence :
* Reverted in all tests of opendj-ldap-sdk the use of a specific LDAPOption to provide a specific class loader, as it is not necessary any more with the fix.
* Removed utility methods in TestCaseUtils class to ease the use of custom class loader, they're not needed anymore

git-svn-id: https://svn.forgerock.org/opendj/trunk@9686 41b1ffd8-f28e-4786-ab96-9950f0a78031