OPENDJ-1217 Servers not synchronized after a conflicting replace on multi-valued attribute on 2 different servers
Fixed regression introduced in r9788 by introducing a similar behaviour as what the code was doing before.

Replica 1 was trying to read a change done on itself and send it to replica 2, but failed to do so.
Hence replica 2 did not receive this UpdateMsg and as a consequence could not resolve it.

Previous code was trying to cursor to a specific CSN in the DB. It could not find it, hence tried to move after it in the DB, but could not do it because the end of the DB was reached. Previous was throwing an exception there, later catching it and then returning an empty cursor.
Newer code removed the useless catch block which let the exception fail this process.
The fix removed throwing the exception and instead returned an empty cursor.



ReplicationDB.java:
In ctor, do not throw an exception, but return an empty cursor instead (Previous code was throwing an exception, catching and ignoring it, then returning an empty cursor).

MessageHandler.java:
Prevented a NullPointerException from being thrown, then caught and ignored.


git-svn-id: https://svn.forgerock.org/opendj/trunk@9887 41b1ffd8-f28e-4786-ab96-9950f0a78031