OPENDJ-1106 fractional-replication: adding entry on a replica as DM should not exit with RC 50
Code review: Matthew Swift


Regression caused by the fix for OPENDJ-948 unauthorized disclosure of directory contents.
The code was looking for the entry matching the DN we are adding which obviously was null here.
Since it was null, the code went on to wrongly check the following conditions: op.isInternalOperation() || op.isSynchronizationOperation() || op.isInnerOperation().
The correct fix was to avoid calling AccessControlHandler.canDiscloseInformation() where it did not make sense at all, when there was no info to protect yet.


LocalBackend*Operation.java:
Do not call the setResultCodeAndMessageNoInfoDisclosure() when we fail to grab a lock, because this is not disclosing any sensitive information at that point.

LocalBackendAddOperation.java:
Do not call the setResultCodeAndMessageNoInfoDisclosure() when we fail to grab a lock, because this is not disclosing any sensitive information at that point.
In lockParent(), reversed the exception to make the code easier to read.

LocalBackendModifyDNOperation.java:
Do not call the setResultCodeAndMessageNoInfoDisclosure() when we fail to grab a lock, because this is not disclosing any sensitive information at that point.
In newDirectoryException(), removed the entryDN parameter since now this method is always called with a non null entry.


AciHandler.java:
Code cleanup


git-svn-id: https://svn.forgerock.org/opendj/trunk@9943 41b1ffd8-f28e-4786-ab96-9950f0a78031