OPENDJ-1482 : SEVERE_ERROR (cursor still opened on the log) when running replication topology with file-based changelog
The root cause is that records may be appended more than once to the 
changelog and in an order that is breaking the key ordering. 

Update Log.java to reject any append of a record that has a key that would
break the ordering of records in the log (ie, key lower than or equal to the last
key appended).

git-svn-id: https://svn.forgerock.org/opendj/trunk@10765 41b1ffd8-f28e-4786-ab96-9950f0a78031